Extracts cmr related operation from ingest/granule.js
Refactoring to ease in moving the cmr function from @cumulus/ingest/granule into @cumulus/cmrjs.
