+++ tasks/post-to-cmr/index.js
@@ -5,7 +5,7 @@ const get = require('lodash.get');
 const cumulusMessageAdapter = require('@cumulus/cumulus-message-adapter-js');
 const { justLocalRun } = require('@cumulus/common/local-helpers');
 const { getCmrFiles } = require('@cumulus/ingest/granule');
-const { publish } = require('@cumulus/ingest/cmr');
+const { publish } = require('@cumulus/ingest/granule-extract-cmr');
 const log = require('@cumulus/common/log');
 const { loadJSONTestData } = require('@cumulus/test-data');
